Major Refactor, clean up UI code. Fix imports to be a pseudo-action
