+++ src/components/rewards/AddRewardPopover.vue
