+++ src/views/TriggersEditor.vue
@@ -2,9 +2,22 @@
   <v-container fluid>
     <v-row style="padding-bottom: 3.5rem">
       <v-col>
-        <trigger-editor v-model="triggers" :triggerKey="triggersName" />
+        <v-row v-for="(commandKey, i) in commands" :key="i">
+          <v-col>
+            <command-card
+              :value="triggers[commandKey]"
+              :actionKey="commandKey"
+              @input="(newData) => updateCommand(commandKey, newData)"
+              @delete="deleteCommand(commandKey)"
+              @key-change="(v) => updateCommandKey(commandKey, v)"
+            />
+          </v-col>
+        </v-row>
       </v-col>
     </v-row>
+    <v-card-actions>
+      <v-btn @click="addCommand"> Add Command </v-btn>
+    </v-card-actions>
     <v-speed-dial v-model="fab" fixed bottom right open-on-hover>
       <template v-slot:activator>
         <v-btn v-model="fab" color="primary" fab>
@@ -27,19 +40,26 @@
 </template>
 
 <script>
-import TriggerEditor from "../components/profiles/TriggerEditor.vue";
+import CommandCard from "../components/commands/CommandCard.vue";
 import YAML from "yaml";
 import fs from "fs";
 import path from "path";
 import { mapGetters } from "vuex";
+import { changeObjectKey } from "../utils/objects.js";
 
 export default {
   components: {
-    TriggerEditor,
+    CommandCard,
     ConfirmDialog: () => import("../components/dialogs/ConfirmDialog.vue"),
   },
   computed: {
     ...mapGetters("ipc", ["paths"]),
+    commands() {
+      if (!this.triggers) {
+        return [];
+      }
+      return Object.keys(this.triggers).filter((key) => key != "imports");
+    },
     triggersName() {
       return this.$route.params.triggers;
     },
@@ -52,6 +72,19 @@ export default {
     };
   },
   methods: {
+    updateCommandKey(oldKey, newKey) {
+      this.triggers = changeObjectKey(this.triggers, oldKey, newKey);
+    },
+    updateCommand(command, data) {
+      this.triggers[command] = data;
+    },
+    deleteCommand(command) {
+      delete this.triggers[command];
+    },
+    addCommand() {
+      this.triggers[""] = { actions: [], sync: false };
+    },
+
     async save() {
       let newYaml = YAML.stringify(this.triggers);
